requirements.txt
aiohttp==3.8.6
aiosqlite==0.20.0
discord.py==2.4.0
python-dotenv==1.0.1
aiofiles==24.1.0
